assignment4/requirements.txt
dnspython==2.7.0
psycopg==3.2.3
psycopg-binary==3.2.3
pymongo==4.10.1
python-dotenv==1.0.1
typing_extensions==4.12.2
